refactor: split lifecycle_helpers/setup_helpers/hybrid_helpers into focused packages (A8, #1888)
