--- src/synthorg/api/lifecycle_helpers.py
@@ -1,1096 +1,0 @@
-"""Smaller lifecycle helpers used by :mod:`synthorg.api.lifecycle_builder`.
-
-Split out of ``lifecycle_builder.py`` so neither file exceeds the
-800-line size budget. Consumers should import from
-``lifecycle_builder`` which re-exports these names.
-"""
-
-import asyncio
-import inspect
-from typing import TYPE_CHECKING, Any
-
-from synthorg.notifications.factory import build_notification_dispatcher
-from synthorg.observability import get_logger, safe_error_description
-from synthorg.observability.events.api import (
-    API_APP_STARTUP,
-    API_AUDIT_RETENTION,
-    API_AUTH_LOCKOUT_CLEANUP,
-    API_BRIDGE_CONFIG_RESOLVE_FAILED,
-    API_SESSION_CLEANUP,
-    API_WS_TICKET_CLEANUP,
-)
-from synthorg.observability.events.persistence import (
-    PERSISTENCE_OAUTH_STATE_CLEANUP,
-)
-from synthorg.observability.events.setup import SETUP_AGENT_BOOTSTRAP_FAILED
-from synthorg.settings.dispatcher import SettingsChangeDispatcher
-from synthorg.settings.enums import SettingNamespace
-from synthorg.settings.registry import (
-    registered_default_bool,
-    registered_default_float,
-    registered_default_int,
-)
-from synthorg.settings.subscribers import (
-    ApiBridgeSettingsSubscriber,
-    BackupSettingsSubscriber,
-    MemorySettingsSubscriber,
-    ObservabilitySettingsSubscriber,
-    PerOpRateLimitSettingsSubscriber,
-    ProviderSettingsSubscriber,
-    SecurityTimeoutSettingsSubscriber,
-)
-
-if TYPE_CHECKING:
-    from collections.abc import Awaitable, Callable
-
-    from synthorg.api.state import AppState
-    from synthorg.backup.service import BackupService
-    from synthorg.communication.bus_protocol import MessageBus
-    from synthorg.config.schema import RootConfig
-    from synthorg.security.timeout.scheduler import ApprovalTimeoutScheduler
-    from synthorg.settings.bridge_configs import NotificationsBridgeConfig
-    from synthorg.settings.service import SettingsService
-    from synthorg.settings.subscriber import SettingsSubscriber
-
-logger = get_logger(__name__)
-
-
-async def _resolve_ticket_cleanup_interval(app_state: AppState) -> float:
-    """Resolve the ticket cleanup interval.
-
-    A settings-backend outage, missing setting, or malformed value must
-    not kill the cleanup task -- otherwise expired WS tickets and
-    sessions accumulate indefinitely until the next restart. Any
-    resolver failure is logged and the registered default for
-    ``api.ticket_cleanup_interval_seconds`` is returned, so the fallback
-    tracks the registry rather than duplicating the literal here.
-    """
-    fallback = registered_default_float(
-        SettingNamespace.API.value, "ticket_cleanup_interval_seconds"
-    )
-    if not app_state.has_config_resolver:
-        return fallback
-    try:
-        return await app_state.config_resolver.get_float(
-            SettingNamespace.API.value, "ticket_cleanup_interval_seconds"
-        )
-    except asyncio.CancelledError:
-        raise
-    except MemoryError, RecursionError:
-        raise
-    except Exception as exc:
-        logger.warning(
-            API_WS_TICKET_CLEANUP,
-            error_type=type(exc).__name__,
-            error=safe_error_description(exc),
-            fallback_seconds=fallback,
-        )
-        return fallback
-
-
-async def _resolve_lifecycle_cleanup_enabled(app_state: AppState) -> bool:
-    """Resolve the lifecycle-cleanup kill-switch.
-
-    Operators flip ``api.lifecycle_cleanup_enabled=false`` to pause the
-    WS ticket / session / lockout cleanup loop mid-flight without tearing
-    down the lifespan task.  A settings-backend outage must not mask the
-    operator's intent in either direction -- the registered default
-    ("keep cleaning") wins because stale tickets and sessions accumulate
-    forever otherwise.
-    """
-    fallback = registered_default_bool(
-        SettingNamespace.API.value, "lifecycle_cleanup_enabled"
-    )
-    if not app_state.has_config_resolver:
-        return fallback
-    try:
-        return await app_state.config_resolver.get_bool(
-            SettingNamespace.API.value, "lifecycle_cleanup_enabled"
-        )
-    except asyncio.CancelledError:
-        raise
-    except MemoryError, RecursionError:
-        raise
-    except Exception as exc:
-        logger.warning(
-            API_WS_TICKET_CLEANUP,
-            error_type=type(exc).__name__,
-            error=safe_error_description(exc),
-            fallback_enabled=fallback,
-        )
-        return fallback
-
-
-async def _run_cleanup_step(
-    action: Callable[[], Awaitable[Any] | Any],
-    *,
-    event: str,
-    failure_message: str,
-) -> None:
-    """Run *action* and log any non-system exception under *event*.
-
-    ``action`` may return either an awaitable (async cleanup) or a
-    synchronous result (e.g. ``ticket_store.cleanup_expired()``). The
-    caller decides whether to gate the call on a ``has_*`` predicate
-    before invoking the helper, keeping the helper's surface
-    deliberately narrow.
-
-    ``MemoryError`` / ``RecursionError`` propagate so the parent loop
-    can crash on truly fatal failures rather than masking them as a
-    routine cleanup blip.
-    """
-    try:
-        result = action()
-        # Use ``inspect.isawaitable`` so Tasks, Futures, and any
-        # ``__await__``-implementing custom awaitable returned by a
-        # cleanup hook are awaited too -- ``asyncio.iscoroutine``
-        # only matches bare coroutines and would silently skip the
-        # rest.
-        if inspect.isawaitable(result):
-            await result
-    except MemoryError, RecursionError:
-        raise
-    except Exception as exc:
-        logger.warning(
-            event,
-            failure_context=failure_message,
-            error_type=type(exc).__name__,
-            error=safe_error_description(exc),
-        )
-
-
-async def _run_cleanup_tick(app_state: AppState) -> None:
-    """Run one cleanup-cycle tick across every store.
-
-    Covers WS tickets, sessions, lockouts, and idempotency keys. Each
-    store's cleanup is delegated to :func:`_run_cleanup_step` so a
-    transient failure in one store cannot abort the others.
-    """
-    from synthorg.observability.events.idempotency import (  # noqa: PLC0415
-        IDEMPOTENCY_CLEANUP,
-    )
-
-    await _run_cleanup_step(
-        app_state.ticket_store.cleanup_expired,
-        event=API_WS_TICKET_CLEANUP,
-        failure_message="Periodic ticket cleanup failed",
-    )
-    if app_state.has_session_store:
-        await _run_cleanup_step(
-            app_state.session_store.cleanup_expired,
-            event=API_SESSION_CLEANUP,
-            failure_message="Periodic session cleanup failed",
-        )
-    if app_state.has_lockout_store:
-        await _run_cleanup_step(
-            app_state.lockout_store.cleanup_expired,
-            event=API_AUTH_LOCKOUT_CLEANUP,
-            failure_message="Periodic lockout cleanup failed",
-        )
-    if app_state.has_persistence:
-        # OAuth state cleanup is invoked directly (not via _run_cleanup_step)
-        # so we can surface the per-tick removed-row count at the lifecycle
-        # observability boundary -- the repo-level log only fires when
-        # ``removed > 0`` (silent on no-op), so a lifecycle log gives
-        # operators an unconditional "sweep ran" signal.
-        oauth_retention_seconds = await _resolve_oauth_idempotency_retention(
-            app_state,
-        )
-        try:
-            oauth_removed = await app_state.persistence.oauth_states.cleanup_expired(
-                oauth_retention_seconds,
-            )
-        except MemoryError, RecursionError:
-            raise
-        except Exception as exc:
-            logger.warning(
-                PERSISTENCE_OAUTH_STATE_CLEANUP,
-                error_type=type(exc).__name__,
-                error=safe_error_description(exc),
-            )
-        else:
-            logger.info(
-                PERSISTENCE_OAUTH_STATE_CLEANUP,
-                note="oauth state sweep completed",
-                removed=oauth_removed,
-            )
-        await _run_cleanup_step(
-            app_state.idempotency_service.cleanup_expired,
-            event=IDEMPOTENCY_CLEANUP,
-            failure_message="Periodic idempotency cleanup failed",
-        )
-
-
-async def _resolve_event_stream_janitor_settings(
-    app_state: AppState,
-) -> tuple[float, float]:
-    """Return ``(idle_ttl, janitor_interval)`` for the EventStreamHub janitor.
-
-    Falls back to the registered defaults when the settings resolver is
-    unavailable or either read fails. The fallback keeps the janitor
-    enabled rather than disabling pruning on a broken settings backend
-    -- leaking subscriber state silently is the worse failure mode.
-    """
-    fallback_idle = registered_default_float(
-        SettingNamespace.COMMUNICATION.value,
-        "event_stream_subscriber_idle_ttl_seconds",
-    )
-    fallback_interval = registered_default_float(
-        SettingNamespace.COMMUNICATION.value,
-        "event_stream_janitor_interval_seconds",
-    )
-    if not app_state.has_config_resolver:
-        return fallback_idle, fallback_interval
-    try:
-        idle = await app_state.config_resolver.get_float(
-            SettingNamespace.COMMUNICATION.value,
-            "event_stream_subscriber_idle_ttl_seconds",
-        )
-        interval = await app_state.config_resolver.get_float(
-            SettingNamespace.COMMUNICATION.value,
-            "event_stream_janitor_interval_seconds",
-        )
-    except asyncio.CancelledError:
-        raise
-    except MemoryError, RecursionError:
-        raise
-    except Exception as exc:
-        logger.warning(
-            API_APP_STARTUP,
-            error_type=type(exc).__name__,
-            error=safe_error_description(exc),
-            fallback_idle_ttl_seconds=fallback_idle,
-            fallback_interval_seconds=fallback_interval,
-        )
-        return fallback_idle, fallback_interval
-    return idle, interval
-
-
-async def _ticket_cleanup_loop(app_state: AppState) -> None:
-    """Periodically prune expired WS tickets and sessions.
-
-    Gated by ``api.lifecycle_cleanup_enabled`` (live, per-tick): when
-    the setting is ``False`` every tick short-circuits -- the loop
-    keeps running so operators can re-enable without restarting, but
-    no cleanup work is done.
-    """
-    while True:
-        await asyncio.sleep(await _resolve_ticket_cleanup_interval(app_state))
-        if not await _resolve_lifecycle_cleanup_enabled(app_state):
-            logger.debug(API_WS_TICKET_CLEANUP, reason="paused_by_setting")
-            continue
-        await _run_cleanup_tick(app_state)
-
-
-async def _resolve_audit_retention_loop_enabled(app_state: AppState) -> bool:
-    """Return whether the audit retention loop is enabled (live, per-tick).
-
-    Reads ``security.audit_retention_loop_enabled`` from the settings
-    resolver.  Falls back to the registered default (``True``) when the
-    resolver is unavailable or the read fails -- leaving expired audit
-    rows around is a compliance risk, so the loop stays active on a
-    broken settings backend rather than silently disabling itself.
-    """
-    fallback = registered_default_bool(
-        SettingNamespace.SECURITY.value, "audit_retention_loop_enabled"
-    )
-    if not app_state.has_config_resolver:
-        return fallback
-    try:
-        return await app_state.config_resolver.get_bool(
-            SettingNamespace.SECURITY.value, "audit_retention_loop_enabled"
-        )
-    except asyncio.CancelledError:
-        raise
-    except MemoryError, RecursionError:
-        raise
-    except Exception as exc:
-        logger.warning(
-            API_AUDIT_RETENTION,
-            error_type=type(exc).__name__,
-            error=safe_error_description(exc),
-            fallback_enabled=fallback,
-        )
-        return fallback
-
-
-async def _resolve_audit_retention_days(app_state: AppState) -> int:
-    """Resolve ``audit_retention_days`` for the retention loop.
-
-    Falls back to the registered default when the settings resolver is
-    unavailable or the read fails.  The fallback intentionally keeps
-    retention enabled rather than disabling purging on a broken
-    settings backend -- leaving expired audit rows around is a
-    compliance risk, so prefer the built-in default to a silent zero.
-    ``0`` is reserved for an operator explicitly opting out via
-    ``security.audit_retention_days=0``.
-    """
-    fallback = registered_default_int(
-        SettingNamespace.SECURITY.value, "audit_retention_days"
-    )
-    if not app_state.has_config_resolver:
-        return fallback
-    try:
-        return await app_state.config_resolver.get_int(
-            SettingNamespace.SECURITY.value, "audit_retention_days"
-        )
-    except asyncio.CancelledError:
-        raise
-    except MemoryError, RecursionError:
-        raise
-    except Exception as exc:
-        logger.warning(
-            API_AUDIT_RETENTION,
-            error_type=type(exc).__name__,
-            error=safe_error_description(exc),
-            fallback_days=fallback,
-        )
-        return fallback
-
-
-async def _resolve_audit_retention_tick_seconds(app_state: AppState) -> float:
-    """Resolve the cadence between audit retention purge ticks.
-
-    Falls back to the registered default when the resolver is
-    unavailable or the read fails.
-    """
-    fallback = registered_default_float(
-        SettingNamespace.SECURITY.value, "audit_retention_tick_seconds"
-    )
-    if not app_state.has_config_resolver:
-        return fallback
-    try:
-        return await app_state.config_resolver.get_float(
-            SettingNamespace.SECURITY.value, "audit_retention_tick_seconds"
-        )
-    except asyncio.CancelledError:
-        raise
-    except MemoryError, RecursionError:
-        raise
-    except Exception as exc:
-        logger.warning(
-            API_AUDIT_RETENTION,
-            error_type=type(exc).__name__,
-            error=safe_error_description(exc),
-            fallback_seconds=fallback,
-        )
-        return fallback
-
-
-async def _audit_retention_tick(app_state: AppState) -> None:
-    """Single iteration of the audit retention sweep.
-
-    Extracted from ``_audit_retention_loop`` so the loop body stays
-    under the project function-length limit.
-    """
-    from datetime import UTC, datetime, timedelta  # noqa: PLC0415
-
-    days = await _resolve_audit_retention_days(app_state)
-    if days <= 0:
-        logger.debug(API_AUDIT_RETENTION, note="audit retention purge disabled")
-        return
-    if not app_state.has_persistence:
-        return
-    cutoff = datetime.now(UTC) - timedelta(days=days)
-    try:
-        deleted = await app_state.persistence.audit_entries.purge_before(cutoff)
-    except MemoryError, RecursionError:
-        raise
-    except Exception as exc:
-        logger.warning(
-            API_AUDIT_RETENTION,
-            note="audit retention purge failed",
-            error_type=type(exc).__name__,
-            error=safe_error_description(exc),
-        )
-        return
-    logger.info(
-        API_AUDIT_RETENTION,
-        note="audit retention purge completed",
-        deleted=deleted,
-        retention_days=days,
-        cutoff=cutoff.isoformat(),
-    )
-
-
-async def _audit_retention_loop(app_state: AppState) -> None:
-    """Daily sweep that purges audit_entries older than retention window.
-
-    Reads ``security.audit_retention_days`` and
-    ``security.audit_retention_loop_enabled`` from the settings
-    resolver on every tick so operator changes take effect without
-    restart.  A ``retention_days`` of 0 disables purging entirely
-    (opt-out via ``security.audit_retention_days=0``); the kill-switch
-    keeps the loop resident but inert so plumbing is unchanged when
-    operators pause retention during incident investigation.
-    Resolver outages fall back to the registered defaults rather than
-    disabling retention -- leaving expired audit rows around is a
-    compliance risk.  Tick cadence comes from
-    ``security.audit_retention_tick_seconds`` (default 24h).
-    """
-    while True:
-        if await _resolve_audit_retention_loop_enabled(app_state):
-            await _audit_retention_tick(app_state)
-        else:
-            logger.info(
-                API_AUDIT_RETENTION,
-                note="audit retention purge paused",
-                reason="paused_by_setting",
-            )
-        await asyncio.sleep(await _resolve_audit_retention_tick_seconds(app_state))
-
-
-async def _maybe_promote_first_owner(app_state: AppState) -> None:
-    """Promote the first user to owner if no owner exists.
-
-    This is a one-time idempotent migration that runs on every boot
-    until at least one user has the ``OrgRole.OWNER`` role.
-    """
-    from datetime import UTC, datetime  # noqa: PLC0415
-
-    if not app_state.has_persistence:
-        return
-    try:
-        users = await app_state.persistence.users.list_users()
-    except asyncio.CancelledError:
-        raise
-    except MemoryError, RecursionError:
-        raise
-    except Exception as exc:
-        logger.warning(
-            API_APP_STARTUP,
-            note="Owner auto-promote skipped: failed to list users",
-            error_type=type(exc).__name__,
-            error=safe_error_description(exc),
-        )
-        return
-    if not users:
-        return
-
-    from synthorg.core.auth.models import OrgRole  # noqa: PLC0415
-
-    has_owner = any(OrgRole.OWNER in u.org_roles for u in users)
-    if has_owner:
-        return
-
-    first = users[0]
-    promoted = first.model_copy(
-        update={
-            "org_roles": (*first.org_roles, OrgRole.OWNER),
-            "updated_at": datetime.now(UTC),
-        },
-    )
-    try:
-        await app_state.persistence.users.save(promoted)
-    except asyncio.CancelledError:
-        raise
-    except MemoryError, RecursionError:
-        raise
-    except Exception as exc:
-        logger.warning(
-            API_APP_STARTUP,
-            note="Owner auto-promote failed",
-            user_id=first.id,
-            error_type=type(exc).__name__,
-            error=safe_error_description(exc),
-        )
-        return
-    logger.info(
-        API_APP_STARTUP,
-        note="Auto-promoted first user to owner",
-        user_id=first.id,
-        username=first.username,
-    )
-
-
-async def _maybe_bootstrap_agents(app_state: AppState) -> None:
-    """Bootstrap agents if setup is complete and services are available.
-
-    On first run, setup isn't complete yet so bootstrap is deferred
-    to ``POST /setup/complete``.  On subsequent starts, agents are
-    loaded from persisted config into the runtime registry.
-    """
-    if not (
-        app_state.has_config_resolver
-        and app_state.has_agent_registry
-        and app_state.has_settings_service
-    ):
-        logger.debug(
-            API_APP_STARTUP,
-            note="Agent bootstrap skipped: required services not available",
-        )
-        return
-
-    try:
-        setup_entry = await app_state.settings_service.get_entry(
-            "api",
-            "setup_complete",
-        )
-        is_complete = setup_entry.value == "true"
-    except asyncio.CancelledError:
-        raise
-    except MemoryError, RecursionError:
-        raise
-    except Exception as exc:
-        logger.warning(
-            API_APP_STARTUP,
-            note="Could not read setup_complete setting; skipping agent bootstrap",
-            namespace="api",
-            key="setup_complete",
-            error_type=type(exc).__name__,
-            error=safe_error_description(exc),
-        )
-        is_complete = False
-
-    if not is_complete:
-        logger.debug(
-            API_APP_STARTUP,
-            note="Agent bootstrap skipped: setup not complete",
-        )
-        return
-
-    try:
-        from synthorg.api.bootstrap import bootstrap_agents  # noqa: PLC0415
-
-        await bootstrap_agents(
-            config_resolver=app_state.config_resolver,
-            agent_registry=app_state.agent_registry,
-        )
-    except asyncio.CancelledError:
-        raise
-    except MemoryError, RecursionError:
-        raise
-    except Exception as exc:
-        logger.warning(
-            SETUP_AGENT_BOOTSTRAP_FAILED,
-            error_type=type(exc).__name__,
-            error=safe_error_description(exc),
-        )
-
-
-def _build_settings_dispatcher(  # noqa: PLR0913 -- one optional arg per subscriber the dispatcher carries
-    message_bus: MessageBus | None,
-    settings_service: SettingsService | None,
-    config: RootConfig,
-    app_state: AppState,
-    backup_service: BackupService | None = None,
-    approval_timeout_scheduler: ApprovalTimeoutScheduler | None = None,
-) -> SettingsChangeDispatcher | None:
-    """Create settings change dispatcher if bus and settings are available."""
-    if message_bus is None or settings_service is None:
-        return None
-    provider_sub = ProviderSettingsSubscriber(
-        config=config,
-        app_state=app_state,
-        settings_service=settings_service,
-    )
-    memory_sub = MemorySettingsSubscriber()
-    log_dir = config.logging.log_dir if config.logging is not None else "logs"
-    observability_sub = ObservabilitySettingsSubscriber(
-        settings_service=settings_service,
-        log_dir=log_dir,
-    )
-    per_op_rl_sub = PerOpRateLimitSettingsSubscriber(
-        app_state=app_state,
-        settings_service=settings_service,
-    )
-    api_bridge_sub = ApiBridgeSettingsSubscriber(
-        app_state=app_state,
-        settings_service=settings_service,
-    )
-    subs: list[SettingsSubscriber] = [
-        provider_sub,
-        memory_sub,
-        observability_sub,
-        per_op_rl_sub,
-        api_bridge_sub,
-    ]
-    if backup_service is not None:
-        subs.append(
-            BackupSettingsSubscriber(
-                backup_service=backup_service,
-                settings_service=settings_service,
-            ),
-        )
-    if approval_timeout_scheduler is not None:
-        subs.append(
-            SecurityTimeoutSettingsSubscriber(
-                scheduler=approval_timeout_scheduler,
-                settings_service=settings_service,
-            ),
-        )
-    config_resolver = (
-        app_state.config_resolver if app_state.has_config_resolver else None
-    )
-    return SettingsChangeDispatcher(
-        message_bus=message_bus,
-        subscribers=tuple(subs),
-        config_resolver=config_resolver,
-    )
-
-
-async def _validate_approval_urgency_invariant(app_state: AppState) -> None:
-    """Reject startup when approval urgency thresholds violate the contract.
-
-    ``api.approval_urgency_critical_seconds`` must be strictly less than
-    ``api.approval_urgency_high_seconds`` -- a critical escalation has
-    to fire sooner than a high one. Both settings are ``restart_required``,
-    so the only place to enforce the cross-setting invariant is at app
-    startup. Registry defaults (3600 / 14400) satisfy the invariant;
-    this guard catches operator-tuned misconfigurations that the
-    per-setting ``min_value`` / ``max_value`` bounds can't express.
-
-    Resolver failures (settings backend down) are logged and the
-    invariant check is skipped -- other bridge-config paths handle the
-    outage independently and the built-in defaults stay safe.
-    """
-    try:
-        critical = await app_state.config_resolver.get_float(
-            SettingNamespace.API.value, "approval_urgency_critical_seconds"
-        )
-        high = await app_state.config_resolver.get_float(
-            SettingNamespace.API.value, "approval_urgency_high_seconds"
-        )
-    except asyncio.CancelledError:
-        raise
-    except MemoryError, RecursionError:
-        raise
-    except Exception as exc:
-        logger.warning(
-            API_APP_STARTUP,
-            error_type=type(exc).__name__,
-            error=safe_error_description(exc),
-        )
-        return
-    if critical >= high:
-        msg = (
-            "Invalid approval-urgency configuration:"
-            f" api.approval_urgency_critical_seconds={critical}"
-            f" must be strictly less than"
-            f" api.approval_urgency_high_seconds={high}."
-            " A critical escalation must fire sooner than a high one."
-        )
-        logger.error(
-            API_APP_STARTUP,
-            error=msg,
-            critical_seconds=critical,
-            high_seconds=high,
-        )
-        raise ValueError(msg)
-
-
-async def _apply_sandbox_image_cache(app_state: AppState) -> None:
-    """Populate the sandbox / sidecar image-resolution cache from settings.
-
-    Called once per startup so ``DockerSandboxConfig`` field defaults
-    stop reading ``os.environ`` directly. ``env_var_override`` on the
-    registered settings preserves the historical
-    ``SYNTHORG_SANDBOX_IMAGE`` / ``SYNTHORG_SIDECAR_IMAGE`` workflow
-    without bypassing the canonical DB > env > YAML > default chain.
-
-    Resolver failures clear the cache to ``None`` so the field default
-    falls through to the documented constant; whitespace-only resolver
-    results are normalised to ``None`` in the caller (the setter also
-    normalises, but stripping here makes the intent explicit).
-    """
-    from synthorg.tools.sandbox._image_resolution import (  # noqa: PLC0415
-        set_resolved_sandbox_image,
-        set_resolved_sidecar_image,
-    )
-
-    for setting_key, setter in (
-        ("sandbox_image", set_resolved_sandbox_image),
-        ("sidecar_image", set_resolved_sidecar_image),
-    ):
-        try:
-            image_value = await app_state.config_resolver.get_str(
-                SettingNamespace.TOOLS.value, setting_key
-            )
-        except asyncio.CancelledError:
-            raise
-        except MemoryError, RecursionError:
-            raise
-        except Exception as exc:
-            setter(None)
-            logger.warning(
-                API_APP_STARTUP,
-                setting=f"tools.{setting_key}",
-                error_type=type(exc).__name__,
-                error=safe_error_description(exc),
-            )
-        else:
-            stripped = image_value.strip() if image_value is not None else None
-            setter(stripped or None)
-
-
-async def _apply_notification_dispatcher_config(
-    app_state: AppState,
-    effective_config: RootConfig | None,
-) -> None:
-    """Rebuild the notification dispatcher with operator-tuned timeouts.
-
-    Reads the notifications bridge config from the resolver, then if a
-    dispatcher already exists on ``app_state``, builds a fresh one with
-    the resolved timeouts and swaps it in. Closes the previous
-    dispatcher's sinks after the swap. Resolver outage falls through to
-    a rebuild with ``bridge_config=None`` (built-in default timeouts) so
-    the live dispatcher still picks up the ``config_resolver`` and the
-    runtime kill-switch stays operational.
-    """
-    notif_bridge: NotificationsBridgeConfig | None
-    try:
-        notif_bridge = await app_state.config_resolver.get_notifications_bridge_config()
-    except asyncio.CancelledError:
-        raise
-    except MemoryError, RecursionError:
-        raise
-    except Exception as exc:
-        logger.warning(
-            API_APP_STARTUP,
-            setting="notifications.bridge_config",
-            error_type=type(exc).__name__,
-            error=safe_error_description(exc),
-        )
-        notif_bridge = None
-    if not (app_state.has_notification_dispatcher and effective_config is not None):
-        return
-    new_dispatcher = build_notification_dispatcher(
-        effective_config.notifications,
-        bridge_config=notif_bridge,
-        config_resolver=app_state.config_resolver,
-    )
-    old_dispatcher = app_state.swap_notification_dispatcher(new_dispatcher)
-    if old_dispatcher is None:
-        return
-    try:
-        await old_dispatcher.aclose()
-    except MemoryError, RecursionError:
-        raise
-    except Exception as exc:
-        logger.warning(
-            API_APP_STARTUP,
-            event_context="old_notification_dispatcher_aclose",
-            error_type=type(exc).__name__,
-            error=safe_error_description(exc),
-        )
-
-
-async def _apply_api_bridge_config_snapshot(app_state: AppState) -> None:
-    """Snapshot ``ApiBridgeConfig`` onto ``AppState`` at startup.
-
-    Resolves the full bridge once via
-    :meth:`ConfigResolver.get_api_bridge_config` and atomically swaps
-    it onto ``app_state``. On any non-fatal resolve failure the
-    default ``ApiBridgeConfig()`` snapshot installed by
-    ``AppState.__init__`` is retained and a single structured warning
-    is emitted -- the centralised replacement for the per-request
-    log-once fallback the activities controller used to carry inline.
-
-    No-op when no resolver is wired (dev/test rigs that bypass
-    ``create_app``); the default snapshot remains in place.
-    """
-    if not app_state.has_config_resolver:
-        return
-    try:
-        snapshot = await app_state.config_resolver.get_api_bridge_config()
-    except asyncio.CancelledError:
-        raise
-    except MemoryError, RecursionError:
-        raise
-    except Exception as exc:
-        logger.warning(
-            API_BRIDGE_CONFIG_RESOLVE_FAILED,
-            bridge="api",
-            error_type=type(exc).__name__,
-            error=safe_error_description(exc),
-            fallback="module_defaults",
-        )
-        return
-    app_state.swap_api_bridge_config(snapshot)
-
-
-async def _apply_bridge_config(  # noqa: C901, PLR0912, PLR0915
-    app_state: AppState,
-    effective_config: RootConfig | None,
-) -> None:
-    """Apply operator-tuned API bridge settings during startup.
-
-    Idempotent via ``app_state.bridge_config_applied`` so a re-entering
-    Litestar lifespan (shared-app test fixtures, multi-lifespan runs)
-    does not churn httpx/SMTP clients or rebuild the OAuth flow.
-    """
-    if not app_state.has_config_resolver or app_state.bridge_config_applied:
-        return
-
-    await _validate_approval_urgency_invariant(app_state)
-    await _apply_api_bridge_config_snapshot(app_state)
-
-    try:
-        app_state.ticket_store.set_max_pending_per_user(
-            await app_state.config_resolver.get_int(
-                SettingNamespace.API.value,
-                "ws_ticket_max_pending_per_user",
-            )
-        )
-    except MemoryError, RecursionError:
-        raise
-    except Exception as exc:
-        logger.warning(
-            API_APP_STARTUP,
-            setting="api.ws_ticket_max_pending_per_user",
-            error_type=type(exc).__name__,
-            error=safe_error_description(exc),
-        )
-
-    try:
-        app_state.set_ws_auth_timeout_seconds(
-            await app_state.config_resolver.get_float(
-                SettingNamespace.API.value,
-                "ws_auth_timeout_seconds",
-            )
-        )
-    except MemoryError, RecursionError:
-        raise
-    except Exception as exc:
-        logger.warning(
-            API_APP_STARTUP,
-            setting="api.ws_auth_timeout_seconds",
-            error_type=type(exc).__name__,
-            error=safe_error_description(exc),
-        )
-
-    # Each WS DoS-prevention setting resolves and applies independently
-    # so a single failed lookup (e.g. settings backend hiccup on one row)
-    # does NOT prevent the other two from being applied. Each falls back
-    # to its built-in default on failure with a structured warning so ops
-    # can see which knob failed.
-    for setting_key, setter_name in (
-        ("ws_frame_timeout_seconds", "set_ws_frame_timeout_seconds"),
-        ("ws_revalidation_window_seconds", "set_ws_revalidation_window_seconds"),
-        ("ws_revalidation_max_failures", "set_ws_revalidation_max_failures"),
-    ):
-        try:
-            value = await app_state.config_resolver.get_int(
-                SettingNamespace.API.value,
-                setting_key,
-            )
-            getattr(app_state, setter_name)(value)
-        except MemoryError, RecursionError:
-            raise
-        except Exception as exc:
-            logger.warning(
-                API_APP_STARTUP,
-                setting=setting_key,
-                error_type=type(exc).__name__,
-                error=safe_error_description(exc),
-            )
-
-    from synthorg.api.auth.token_size import (  # noqa: PLC0415
-        _DEFAULT_AUTH_TOKEN_BYTES,
-        set_auth_token_bytes,
-    )
-
-    try:
-        set_auth_token_bytes(
-            await app_state.config_resolver.get_int(
-                SettingNamespace.SECURITY.value,
-                "auth_token_bytes",
-            )
-        )
-    except MemoryError, RecursionError:
-        raise
-    except Exception as exc:
-        # Make the safe default *real*: the warning previously only
-        # logged ``fallback_bytes`` without applying it, so a prior
-        # successful resolve that left the cache at a non-default
-        # value would persist after this branch fired.  Force the
-        # cache back to the registered default so the operator-facing
-        # log claim matches process state.
-        set_auth_token_bytes(_DEFAULT_AUTH_TOKEN_BYTES)
-        logger.warning(
-            API_APP_STARTUP,
-            setting="security.auth_token_bytes",
-            error_type=type(exc).__name__,
-            error=safe_error_description(exc),
-            fallback_bytes=_DEFAULT_AUTH_TOKEN_BYTES,
-        )
-
-    from synthorg.engine.timeout_enforcement import (  # noqa: PLC0415
-        set_timeout_enforcement_enabled,
-    )
-
-    try:
-        set_timeout_enforcement_enabled(
-            value=await app_state.config_resolver.get_bool(
-                SettingNamespace.ENGINE.value,
-                "timeout_enforcement_enabled",
-            )
-        )
-    except MemoryError, RecursionError:
-        raise
-    except Exception as exc:
-        # Make the safe default *real*: the warning previously logged
-        # ``fallback_enabled=True`` but did not actually call the
-        # setter, so a misconfigured deployment whose resolver had
-        # already returned False on a prior request would keep
-        # enforcement off even after this branch fires.  Force the
-        # cache back to True so the operator-facing log claim
-        # matches process state.
-        set_timeout_enforcement_enabled(value=True)
-        logger.warning(
-            API_APP_STARTUP,
-            setting="engine.timeout_enforcement_enabled",
-            error_type=type(exc).__name__,
-            error=safe_error_description(exc),
-            fallback_enabled=True,
-        )
-
-    await _apply_sandbox_image_cache(app_state)
-
-    if app_state.oauth_token_manager is not None:
-        app_state.oauth_token_manager.set_config_resolver(
-            app_state.config_resolver,
-        )
-    if app_state.webhook_event_bridge is not None:
-        app_state.webhook_event_bridge.set_config_resolver(
-            app_state.config_resolver,
-        )
-    if app_state.escalation_notify_subscriber is not None:
-        app_state.escalation_notify_subscriber.set_config_resolver(
-            app_state.config_resolver,
-        )
-    _bus = app_state.message_bus if app_state.has_message_bus else None
-    if _bus is not None:
-        _set_resolver = getattr(_bus, "set_config_resolver", None)
-        if callable(_set_resolver):
-            _set_resolver(app_state.config_resolver)
-
-    try:
-        signing_timeout = await app_state.config_resolver.get_float(
-            SettingNamespace.OBSERVABILITY.value,
-            "audit_chain_signing_timeout_seconds",
-        )
-    except MemoryError, RecursionError:
-        raise
-    except Exception as exc:
-        logger.warning(
-            API_APP_STARTUP,
-            setting="observability.audit_chain_signing_timeout_seconds",
-            error_type=type(exc).__name__,
-            error=safe_error_description(exc),
-        )
-    else:
-        from synthorg.observability.audit_chain.sink import (  # noqa: PLC0415
-            AuditChainSink,
-        )
-        from synthorg.observability.startup_wiring import (  # noqa: PLC0415
-            _iter_logging_handlers,
-        )
-
-        for _handler in _iter_logging_handlers():
-            if isinstance(_handler, AuditChainSink):
-                try:
-                    _handler.set_signing_timeout_seconds(signing_timeout)
-                except MemoryError, RecursionError:
-                    raise
-                except Exception as exc:
-                    logger.warning(
-                        API_APP_STARTUP,
-                        setting=("observability.audit_chain_signing_timeout_seconds"),
-                        phase="apply_to_handler",
-                        error_type=type(exc).__name__,
-                        error=safe_error_description(exc),
-                    )
-
-    await _apply_notification_dispatcher_config(app_state, effective_config)
-
-    app_state.mark_bridge_config_applied()
-
-
-async def _apply_security_timeout_interval(
-    app_state: AppState,
-    scheduler: ApprovalTimeoutScheduler | None,
-) -> None:
-    """Apply ``security.timeout_check_interval_seconds`` at startup.
-
-    The scheduler is bootstrapped with the registry default at app
-    construction time before persistence connects. Once the resolver
-    is wired (after ``_apply_bridge_config``), pull the operator-tuned
-    value via the canonical DB > env > YAML > default chain and call
-    ``scheduler.reschedule(...)`` so the configured cadence takes
-    effect on the next loop tick.
-
-    Resolver outage falls back to the *current* scheduler interval
-    (the registry default the scheduler was bootstrapped with). The
-    fail-safe-on-outage rule from the kill-switch idiom applies in
-    spirit: leaving the scheduler running with the bootstrap default
-    is safer than stopping it on a settings-backend hiccup.
-    """
-    if scheduler is None or not app_state.has_config_resolver:
-        return
-    fallback = registered_default_float(
-        SettingNamespace.SECURITY.value,
-        "timeout_check_interval_seconds",
-    )
-    try:
-        interval = await app_state.config_resolver.get_float(
-            SettingNamespace.SECURITY.value,
-            "timeout_check_interval_seconds",
-        )
-    except asyncio.CancelledError:
-        raise
-    except MemoryError, RecursionError:
-        raise
-    except Exception as exc:
-        logger.warning(
-            API_APP_STARTUP,
-            setting="security.timeout_check_interval_seconds",
-            error_type=type(exc).__name__,
-            error=safe_error_description(exc),
-            fallback_seconds=fallback,
-        )
-        return
-    try:
-        scheduler.reschedule(interval)
-    except ValueError as exc:
-        logger.warning(
-            API_APP_STARTUP,
-            setting="security.timeout_check_interval_seconds",
-            value=interval,
-            error_type=type(exc).__name__,
-            error=safe_error_description(exc),
-        )
-
-
-async def _resolve_oauth_idempotency_retention(app_state: AppState) -> float:
-    """Resolve the OAuth idempotency retention window.
-
-    Falls back to the registered default when the resolver is
-    unavailable or the read fails. A settings-backend outage must not
-    stop the OAuth state cleanup loop; the table would otherwise grow
-    unbounded as consumed-but-stale rows accumulate.
-    """
-    fallback = registered_default_float(
-        SettingNamespace.INTEGRATIONS.value,
-        "oauth_idempotency_retention_seconds",
-    )
-    if not app_state.has_config_resolver:
-        return fallback
-    try:
-        return await app_state.config_resolver.get_float(
-            SettingNamespace.INTEGRATIONS.value,
-            "oauth_idempotency_retention_seconds",
-        )
-    except asyncio.CancelledError:
-        raise
-    except MemoryError, RecursionError:
-        raise
-    except Exception as exc:
-        logger.warning(
-            PERSISTENCE_OAUTH_STATE_CLEANUP,
-            setting="integrations.oauth_idempotency_retention_seconds",
-            error_type=type(exc).__name__,
-            error=safe_error_description(exc),
-            fallback_seconds=fallback,
-        )
-        return fallback
