+++ tests/unit/api/test_auto_promote.py
@@ -5,7 +5,7 @@ from datetime import UTC, datetime
 import pytest
 
 from synthorg.api.approval_store import ApprovalStore
-from synthorg.api.lifecycle_helpers import _maybe_promote_first_owner
+from synthorg.api.lifecycle_helpers.bootstrap import _maybe_promote_first_owner
 from synthorg.api.state import AppState
 from synthorg.config.schema import RootConfig
 from synthorg.core.auth.models import OrgRole, User
